[*] Refactoring styles
