+++ package.json
@@ -11,8 +11,9 @@
     "development": "npm run next-server",
     "start": "npm run next-server",
     "next-server": "node ./src/server/app.js",
-    "test": "npm run lint",
-    "lint": "eslint --fix ./src",
+    "test": "run-s test:**",
+    "test:eslint": "eslint --fix ./src",
+    "test:lint-styles": "stylelint \"src/client/**/*.pcss\"",
     "add-contributor": "all-contributors add",
     "generate-contributors": "all-contributors generate"
   },
@@ -20,18 +21,21 @@
   "license": "MIT",
   "dependencies": {
     "@paralect/common-logger": "0.1.10",
-    "@zeit/next-css": "0.1.1",
+    "@zeit/next-css": "0.1.2",
     "axios": "0.17.1",
     "classnames": "2.2.5",
     "cssnano": "3.10.0",
     "lodash": "4.17.5",
-    "next": "5.0.0",
+    "next": "5.0.1-canary.2",
+    "npm-run-all": "4.1.2",
     "postcss-cssnext": "3.1.0",
     "postcss-import": "11.1.0",
     "prop-types": "15.6.0",
     "react": "16.2.0",
     "react-dom": "16.2.0",
     "react-icon-base": "2.1.2",
+    "stylelint": "8.4.0",
+    "stylelint-config-standard": "18.0.0",
     "webpack": "3.11.0"
   },
   "devDependencies": {
@@ -41,7 +45,7 @@
     "babel-plugin-inline-react-svg": "0.5.2",
     "babel-plugin-module-resolver": "3.1.0",
     "babel-preset-es2015": "6.24.1",
-    "eslint": "4.17.0",
+    "eslint": "4.18.0",
     "eslint-plugin-import": "2.8.0",
     "eslint-plugin-jsx-a11y": "6.0.3",
     "eslint-plugin-react": "7.6.1",
