+++ src/client/components/button/styles.pcss
@@ -1,10 +1,11 @@
-button {
+.button {
   transition: 0.5s;
   border-radius: 5px;
   cursor: pointer;
   color: white;
   padding: 15px 10px;
   width: 100%;
+  line-height: 22px;
 
   &[disabled] {
     cursor: default;
@@ -13,6 +14,20 @@ button {
   &.loading {
     animation: Gradient 1s ease-in-out infinite;
     background-position: right center;
-    opacity: .8;
+    opacity: 0.8;
+  }
+}
+
+@keyframes Gradient {
+  0% {
+    background-position: 0% 50%;
+  }
+
+  50% {
+    background-position: 100% 50%;
+  }
+
+  100% {
+    background-position: 0% 50%;
   }
 }
