+++ src/client/components/error/styles.pcss
@@ -1,4 +1,4 @@
-@import 'variables.css';
+@import 'variables.pcss';
 
 .error {
   background: color(red alpha(-50%));
