+++ src/client/components/form/index.js
@@ -1,0 +1,5 @@
+import Wrap from './wrap';
+import Form from './form';
+
+export { Wrap };
+export default Form;
