+++ src/client/components/form/styles.pcss
@@ -1,0 +1,30 @@
+@import 'variables.pcss';
+
+.form {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+
+  & input {
+    margin: var(--form-padding) 0;
+  }
+
+  & button[action="submit"] {
+    margin-top: var(--form-padding);
+  }
+
+  & p {
+    display: block;
+    width: 100%;
+  }
+}
+
+.wrap {
+  display: flex;
+  flex-direction: column;
+  padding: 0 30px;
+  justify-content: center;
+  align-items: flex-start;
+  width: 100%;
+  height: 100%;
+}
