+++ src/client/layouts/main/styles.pcss
@@ -1,12 +1,12 @@
-@import 'variables.css';
+@import 'variables.pcss';
+@import 'global.pcss';
 
 .wrap {
   display: flex;
   flex-direction: column;
   min-height: 100vh;
-
   background-color: var(--color-brand);
-  box-shadow: 0 7px 30px 3px rgba(94,96,186,.35);
+  box-shadow: 0 7px 30px 3px rgba(94, 96, 186, 0.35);
 
   &.secondary {
     background-color: var(--color-secondary);
