+++ src/client/pages/_document.jsx
@@ -2,8 +2,6 @@
 import React from 'react';
 import Document, { Head, Main, NextScript } from 'next/document';
 
-import GlobalStyle from '~/styles/global';
-
 export default class MyDocument extends Document {
   render() {
     return (
@@ -13,10 +11,9 @@ export default class MyDocument extends Document {
           <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1" />
           <link rel="shortcut icon" href="static/favicon.ico" />
           <link href="https://fonts.googleapis.com/css?family=Open+Sans:400,600" rel="stylesheet" />
-
+          <link rel="stylesheet" href="/_next/static/style.css" />
         </Head>
         <body>
-          <GlobalStyle />
           <Main />
           <NextScript />
         </body>
