+++ src/client/pages/forgot-password/styles.pcss
@@ -1,0 +1,12 @@
+@import 'variables.pcss';
+
+.panel {
+  width: 500px;
+  height: auto;
+  padding: var(--form-padding);
+}
+
+.submit-btn {
+  margin: var(--form-padding) 0;
+  background: var(--button-primary-gradient);
+}
