+++ src/client/pages/main/styles.pcss
@@ -1,5 +1,4 @@
-@import 'variables.css';
-@import 'container.css';
+@import 'variables.pcss';
 
 .page {
   flex: 1;
@@ -21,7 +20,7 @@
     & .description {
       color: var(--color-white);
       font-size: 20px;
-      opacity: .85;
+      opacity: 0.85;
     }
   }
 
@@ -30,21 +29,19 @@
     max-width: 400px;
     min-width: 300px;
     min-height: 250px;
-
     display: flex;
     justify-content: center;
 
     & div {
-      display: block;
+      display: flex;
       width: 100%;
       flex: 1;
-      background: linear-gradient(to right, rgba(90, 97, 241, .9) 0%, #a900ff 100%);
+      background: linear-gradient(to right, rgba(90, 97, 241, 0.9) 0%, #a900ff 100%);
       border-radius: 50px;
       transform: skewX(-5deg);
       font-size: 100px;
       font-weight: 600;
-      color: rgba(255, 255, 255, .8);
-      display: flex;
+      color: rgba(255, 255, 255, 0.8);
       align-items: center;
       justify-content: center;
     }
@@ -52,10 +49,14 @@
 }
 
 .container {
+  display: flex;
+  flex-flow: row wrap;
   align-items: center;
   flex: 1;
   justify-content: space-between;
   width: 100%;
-  max-width: var(--max-container-width);
   box-sizing: border-box;
+  max-width: var(--max-container-width);
+  padding: var(--container-padding);
+  margin: 0 auto;
 }
