+++ src/client/pages/reset-password/styles.pcss
@@ -1,0 +1,22 @@
+@import 'variables.pcss';
+
+.panel {
+  width: 500px;
+  height: auto;
+  padding: var(--form-padding);
+
+  & .form-wrap {
+    display: flex;
+    flex-direction: column;
+    padding: 0 30px;
+    justify-content: center;
+    align-items: flex-start;
+    width: 100%;
+    height: 100%;
+
+    & button[action="submit"] {
+      margin: var(--form-padding) 0;
+      background: var(--button-primary-gradient);
+    }
+  }
+}
