--- src/client/pages/signin/styles.css
@@ -1,50 +1,0 @@
-@import 'variables.css';
-@import 'form.css';
-@import 'panel.css';
-@import 'auth.css';
-
-.page {
-  background-color: var(--color-brand);
-  @custom-media --navbar-height-reached (height <= 700px);
-
-  @media (--navbar-height-reached) {
-    padding-top: 20px;
-    align-items: flex-start;
-  }
-
-  & .panel {
-    width: 850px;
-
-    & img {
-      min-width: 400px;
-    }
-
-    & .form-wrap {
-      display: flex;
-      flex-direction: column;
-      padding: 0 30px;
-      justify-content: center;
-      align-items: center;
-      width: 100%;
-      height: 100%;
-
-      & form {
-        & .forgot {
-          width: 100%;
-          text-align: right;
-        }
-
-        & :global(button.signin) {
-          background: var(--button-primary-gradient);
-          margin-top: var(--form-padding)
-        }
-
-        & .signup {
-          width: 100%;
-          text-align: center;
-          margin-top: 2rem;
-        }
-      }
-    }
-  }
-}
