+++ src/client/pages/signin/styles.pcss
@@ -1,0 +1,30 @@
+@import 'variables.pcss';
+
+.panel {
+  width: 850px;
+
+  & img {
+    min-width: 400px;
+  }
+}
+
+.title {
+  display: block;
+  text-align: center;
+  width: 100%;
+}
+
+.forgot {
+  width: 100%;
+  text-align: right;
+}
+
+.signin {
+  background: var(--button-primary-gradient);
+}
+
+.signup {
+  width: 100%;
+  text-align: center;
+  margin-top: 2rem;
+}
