+++ src/client/pages/signup/styles.pcss
@@ -1,0 +1,40 @@
+@import 'variables.pcss';
+
+.panel {
+  height: 600px;
+  width: 1014px;
+
+  & img {
+    min-width: 450px;
+  }
+
+  & h2 {
+    width: 100%;
+    text-align: center;
+  }
+
+  & form {
+    & input:first-child {
+      margin-right: 16px;
+    }
+
+    & .names {
+      display: flex;
+
+      & input {
+        width: 50%;
+      }
+    }
+
+    & .signup {
+      background: linear-gradient(to right, #18c76d 0%, #08af81 100%);
+      margin-top: var(--form-padding);
+    }
+  }
+
+  & .signup-success {
+    margin: auto;
+    width: 100%;
+    padding: 30px;
+  }
+}
