--- src/client/styles/auth.css
@@ -1,42 +1,0 @@
-@import 'variables.css';
-
-.auth.page {
-  @custom-media --breakpoint-small (width <= 720px);
-  flex: 1;
-  width: 100%;
-  padding-top: 20px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  overflow: hidden;
-  position: relative;
-  & .panel {
-    height: 600px;
-    @media (--breakpoint-small) {
-      width: 100%;
-    }
-    & img {
-      width: auto;
-      height: 100%;
-      border-radius: 10px;
-      background: var(--color-background);
-      @media(--breakpoint-small) {
-        display: none;
-      }
-    }
-    & .form-wrap {
-      display: flex;
-      flex-direction: column;
-      padding: 0 30px;
-      justify-content: center;
-      align-items: center;
-      width: 100%;
-      height: 100%;
-      & form {
-        & a {
-          color: #a1a1a1;
-        }
-      }
-    }
-  }
-}
