--- src/client/styles/form.css
@@ -1,25 +1,0 @@
-@import 'variables.css';
-
-form {
-  &.form {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    & .input {
-      border-radius: 8px;
-      padding: var(--form-padding) calc(var(--form-padding) * 2);
-      margin: var(--form-padding) 0;
-      background: white;
-      border: 1px solid #e8e8e8;
-      transition: box-shadow .5s;
-      &:focus {
-        outline: none;
-        /*box-shadow: 0px 0px 10px color(var(--color-brand) alpha(-10%));*/
-        box-shadow: 0px 0px 10px color(#55299a alpha(-50%));
-      }
-    }
-    & button[type="submit"] {
-      margin: 16px;
-    }
-  }
-}
