--- src/client/styles/global.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-import './global.css';
-
-export default () => (
-  <div />
-);
