+++ src/client/styles/global.pcss
@@ -1,4 +1,4 @@
-@import 'variables.css';
+@import 'variables.pcss';
 
 * {
   font-size: 100%;
