--- src/client/styles/panel.css
@@ -1,10 +1,0 @@
-@import 'variables.css';
-
-.panel {
-  padding: 5px;
-  display: flex;
-  background: var(--color-white);
-  border-radius: 10px;
-  box-shadow: 0 7px 30px 3px rgba(94,96,186,.35);
-  margin: 10px;
-}
