+++ src/client/styles/variables.pcss
@@ -1,16 +1,11 @@
 :root {
   --bgColor: red;
   --color-white: #fff;
-
   --color-brand: #55299a;
   --color-secondary: #59ca79;
   --color-background: #fafafa;
-
   --form-padding: 16px;
-
   --max-container-width: 1160px;
   --container-padding: 0 20px;
-
-  --button-primary-gradient: linear-gradient(to right, rgba(90,97,241,0.9) 0%, #7a00ff 100%);
+  --button-primary-gradient: linear-gradient(to right, rgba(90, 97, 241, 0.9) 0%, #7a00ff 100%);
 }
-
